fix(connector): support for walletconnect on remix
