+++ app/root.tsx
@@ -27,6 +27,7 @@ export async function loader() {
     ENV: {
       CHAIN_ID: process.env.CHAIN_ID,
       RPC_URL: process.env.RPC_URL,
+      INFURA_ID: process.env.INFURA_ID,
     },
   });
 }
