+++ package.json
@@ -10,12 +10,13 @@
     "dev": "cross-env NODE_ENV=development remix dev",
     "start": "cross-env NODE_ENV=production remix-serve build",
     "sync-assets": "copy-aragon-ui-assets ./public",
-    "postinstall": "remix setup node"
+    "postinstall": "npx patch-package && remix setup node"
   },
   "dependencies": {
     "@1hive/1hive-ui": "^1.0.18",
     "@remix-run/react": "^1.2.3",
     "@remix-run/vercel": "^1.2.3",
+    "buffer": "^6.0.3",
     "clipboard-polyfill": "^4.0.0-rc1",
     "ethers": "^5.6.1",
     "react": "^17.0.2",
@@ -36,6 +37,7 @@
     "@types/stylis": "^4.0.2",
     "cross-env": "^7.0.3",
     "eslint": "^8.9.0",
+    "patch-package": "^6.4.7",
     "typescript": "^4.5.5"
   },
   "engines": {
